requirements.txt
tabulate
fabulous-color
psycopg2
